uml-editor-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
asyncpg
databases[postgresql]
redis